refactor: integrate async/await; add babel + jest
