+++ .gitignore
@@ -4,3 +4,4 @@ yarn-error.log
 package-lock.json
 .vscode
 presentation.pdf
+lib
