+++ package.json
@@ -2,14 +2,17 @@
   "name": "spectacle-renderer",
   "version": "0.0.2",
   "description": "Render Spectacle presentations with Headless Chrome",
-  "bin": "bin/spectacle-renderer.js",
+  "bin": "lib/cli.js",
   "main": "lib/index.js",
   "scripts": {
     "lint": "eslint .",
+    "prebuild": "npm run clean",
+    "build": "babel src --out-dir lib",
+    "clean": "del-cli 'lib/**/*'",
     "start": "node bin/spectacle-renderer.js",
     "precommit": "lint-staged",
-    "preversion": "npm run lint && npm run test",
-    "test": "mocha \"test/**/*.spec.js\" --timeout 10000"
+    "preversion": "npm run lint && npm run test && npm run build",
+    "test": "jest"
   },
   "engines": {
     "node": ">=6.4.0"
@@ -39,13 +42,16 @@
     "yargs": "^9.0.1"
   },
   "devDependencies": {
+    "babel-cli": "~6.26.0",
     "babel-eslint": "^8.0.1",
-    "chai": "^4.1.2",
+    "babel-preset-env": "~1.6.1",
+    "del-cli": "~1.1.0",
     "eslint": "^4.9.0",
     "eslint-config-formidable": "^3.0.0",
     "eslint-plugin-filenames": "^1.2.0",
     "eslint-plugin-import": "^2.7.0",
     "husky": "^0.14.3",
+    "jest": "~21.2.1",
     "lint-staged": "^4.2.3",
     "mocha": "^4.0.1",
     "pify": "^3.0.0",
