--- test/.eslintrc
@@ -1,8 +1,0 @@
----
-extends:
-  - ../.eslintrc
-
-globals:
-  describe: true
-  afterEach: true
-  it: true
